feat: atom 에 있던 interface 분리 및 새로운 interface 추가
